Re-enable compilation of idc for MinGW
Re-enable compilation of idc for MinGW to fulfill the activeqt dependency
on idc.

idc and activeqt compile fine with MinGW, but most examples won't work
because of the missing midl compiler. This was the reason activeqt and idc
was disabled originally after Qt 4.8.4. Anyhow, 330ca76b then re-enabled
compilation for activeqt, but not idc, which broke the build. This patch
re-enables compilation of idc, but still by default does not compile
activeqt examples.

Qt 5 ships with activeqt / idc for MinGW too, btw, so excluding it in Qt
4.8.5 is inconsistent.

Task-number: QTBUG-31517
Change-Id: I7d4e7ed56c1954996b7291f56143ff83fcdcb0a5
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>