Only use Vista style for Wizards when on Vista or later
If the application is running in XP compatibility mode then we don't
want to show the Vista style for the wizard as this will include a back
button that should not exist. Therefore we check the platform it is on
before trying to resolve the symbols so that it is correct.

Change-Id: Ic7249e06461068fe08184a22bb750ab8736ebc3f
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>