QNX: Changing the way JSC retrieves the current stack base pointer
What Qt is calling the stackbase is the top of the stack.
The thread local storage [__tls()] area is at the top of the stack
and the stack pointer is initialized below this on thread
creation.
With this patch, the stack base pointer can be retrieved faster then in
the current implementation.

This patch is from Sean Boudreau.

Change-Id: I3d1ac58d5c43997cbf462424c66be0c7caafcf1b
Reviewed-by: Thomas McGuire <thomas.mcguire@kdab.com>
Reviewed-by: Alan Alpert <aalpert@blackberry.com>
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>