Mac: Handle the maximizing of the window ourselves when it is frameless
On Mac we need to make a frameless window appear maximized manually
rather than letting the underlying Cocoa API deal with it because
it would overwise not appear correctly.

The test is only done for Mac due to the fact that it is not
giving reliable results on other platforms and the source code
change is Mac specific anyway.

Change-Id: Id48a67ba70bfb4bdc921256f1a80328615c98a6b
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>