Correct algorithm for digest auth when using the CONNECT verb
QHttpSocketEngine fails to authenticate to an HTTP proxy that is using
Digest authentication and the CONNECT method (i.e. when you are
tunneling TLS over HTTP). The bug is due to a bad parameter being
passed to QAuthenticatorPrivate::calculateResponse - the requestMethod
parameter is passed in as "CONNECT " instead of "CONNECT" (note the
trailing space).

Because an MD5 hash is derived from this method when using the
qop="auth" flavor of Digest auth, the hash does not match the expected
value and authentication always fails in this configuration.

(cherry picked from commit 5cab14b8a1dfbb03e22b10af385fb90900a9f280)

Change-Id: Ief025ada714e03d96a316116f6b9f1711ab2a7f7
Reviewed-by: Peter Hartmann <phartmann@blackberry.com>