Fix compile when QT_NO_WARNING_OUTPUT defined without QT_NO_DEBUG_STREAM
qWarning is forward declared as returning QDebug when QT_NO_DEBUG_STREAM
is not defined. This conflicts with the declaration of qWarning
returning QNoDebug when QT_NO_WARNING_OUTPUT is defined.

Avoid forward declaration of qWarning when QT_NO_WARNING_OUTPUT is not
defined to fix compilation.

This does not apply to Qt 5.

Task-number: QTBUG-33301
Change-Id: Ia4750cd63ac404700aedc36f07e5516d4163e9d2
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>