Windows: Fix tablet position in relative (mouse) mode.
When in "mouse" or "relative" mode with the pen position information
would not be calculated correctly resulting in a significant offset
between the tablet pen location and the mouse cursor location. Logic
was added to detect when the two were not in sync and use the mouse
location when this happens.

Change-Id: Id3db5f2de0a657a0d072cee95c6b27179ea9182a
Reviewed-by: Laszlo Agocs <laszlo.agocs@digia.com>
(cherry picked from qtbase/17ebcd2b4690f73c8fd2332b0ba55b3ee3e2e8bb)