Fix tab-completion for variables that have digits in them
In the QtScript Debugger if a variable had a number in it then it would
not tab complete to show the member variables. This accounts for digits
so that it is correctly able to find the variable to complete on.

Change-Id: I4fe0b3b776aba0e3caa6dd524c1a0c59421ea515
Reviewed-by: Mitch Curtis <mitch.curtis@digia.com>
(cherry picked from qtscript/925f1264c590c12aaa5f93a687c7b2aba4370115)