Fix problem with QNetworkAddressEntry returning a invalid netmask.
The ip address should be set before the netmask. The reason for this is
that QNetworkAddressEntry::setNetmask() compares the protocol of the
netmask and the ip, if they don't match the netmask won't be set.

Task-number: QTBUG-33911
Change-Id: Ic344b3653c5dfdc5df912dee16e4dbe069d57d24
(cherry-picked from qtbase commit e82951611c6d965cc8ef9cbf4ee7fbccf615bb3f)
Reviewed-by: Peter Hartmann <phartmann@blackberry.com>