QWindowsVistaStyle: Fixed position of menu gutter when using a custom widget action.
The gutter position was hardcoded to 28 which fits when drawing
on the QMenu with a left margin of 3. It became offset when drawing
on a custom widget with no margin. Calculate position relative
to margin for left to right.

[ChangeLog][QtWidgets][QMenu][Windows] Fixed position of menu gutter when using a custom widget action.

Task-number: QTBUG-36218

Change-Id: Ibb34a9fba2bab7257482b80d2e51e76a9755345c
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>
(cherry picked from qtbase/a5c288ea5675a382edf0b3ecae42f461a22a7303)