Ensure the webftp example works with newer versions of webkit.
The code currently has a bug that it will return the incorrect number
for bytesAvailable() since it doesn't take account of the QIODevice
buffer. This fails when peek() is called (which is used by newer
webkits). See QTBUG-27469 for an example of the problem.

Note this example does not exist in qt5 so this is not a backport.

Change-Id: Ic61248d0f36db6a6a65d0c48a8f40346dc7c3cb6
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@digia.com>