Webbrowser example: Fix crash due to deprecated code.
The example uses a Qt Designer form with widgets on a QToolBar
imported from Qt 3 which is not supported in Qt 4 and later.

In the code generated by uic, a QLabel is constructed passing
the toolbar as parent and added to the toolbar using
QToolBar::addWidget() which then causes a crash in Qt 4.

Move the widgets from the form into the class to avoid this.

Task-number: QTBUG-36948

Change-Id: Ia4af871ce99faf555bebc6ebde19bcae0de4d29e
Reviewed-by: Andy Shaw <andy.shaw@digia.com>
(cherry picked from qtactiveqt/2e44efa5056355ac2c069ad2bd8a8ba6562f3eb6)