Fix building of WebKit's JSC for earlier versions of gcc on Mac
Since building for Carbon typically uses older versions of gcc on older
versions of Mac, then we need to revert
e3c44790065894f4e7f98ab097fce22c5bcbbd0a partially in order to fix this.
Since the original code was just a cleanup then this is safe to do so.

Change-Id: Ic662841f53660b763af2bc71cb9b4684fe614063
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>