Assistant: Set the url on created QNetworkReply objects.
WebKit needs this as a base url in order to resolve links in css for
example. If index.html is at the location doc/, and it loads a css
stylesheet at href="_static/first.css", and first.css contains

 @import url("second.css")

Then second.css should be found in _static, relative to first.css.

That only works if WebKit knows the requested url of the loading
document, so make sure that is set on the reply object.

This is a backport of part of commit qttools/e38ee135f3cf74f413dc7a3e4a9f97a6b5d903bf

Task-number: QTCREATORBUG-8211
Change-Id: Ia92519ab37269a6d3f8bf0e4f8b348595f1f3d59
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>