Fix detection of newer MinGW toolchains in configure.exe
This fixes a regression in the new MinGW packages that are now built
with mkspec win32-g++-4.6: Since we failed to detect this as a MinGW
mkspec we missed to check for headers in all relevant paths, which
disabled windowsxp, windosvista styles by default.

This commit does not apply to Qt 5, similar issues there are being
fixed independently.

[ChangeLog][Windows] configure.exe now detects newer MinGW versions
correctly, enabling windowsxp and windowsvista styles by default.

Task-number: QTBUG-38706
Change-Id: Ib4e7af69e8040aaa28940f501cce5572588d4378
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@digia.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>