OSX: avoid crash in invalidateBuffer_resizeHelper on parentless widget
KDE4 has a KMenu::addTitle function that adds a distinctive item to
a menu, as the predecessor to Qt5's QMenu::addSection. On OS X this
function leads to a null pointer access because
invalidateBuffer_resizeHelper is called on a parentless widget.
This patch filters out such widgets.

Task-number: QTBUG-41716
Change-Id: I8f62a14d201ce44e038ceff87ee5db98277f1b87
Reviewed-by: Morten Johan Sørvig <morten.sorvig@theqtcompany.com>