QGLWidget and re-parenting
On OS X we have QWidgetPrivate::glWidgets list to track all child gl-widgets
and this propagates though all the widget's tree, starting from QGLWidget and up to the
top-level window ('isWindow'). It can happen, that some of widgets in a chain is becoming
a window (but still has a parent) - this results in QGLWidget's destructor failing to remove
'this' from one of ancestors (since it stops on the first 'isWindow' ancestor).
Modify QWidgetPrivate::setParent_sys that it does NOT add glWidgets into the parent's
glWidgets, if this widget isWindow itself.

Change-Id: I4bf4a927a4e43537355fdcdef465638c91e9f956
Task-number: QTBUG-20652
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@theqtcompany.com>